8 refactor styles
