+++ client/src/containers/App/Component.js
@@ -1,33 +1,17 @@
 import PropTypes from 'prop-types'
 import React from 'react'
-import { Scrollbars as S } from 'react-custom-scrollbars'
 import ReactGA from 'react-ga'
 import { Helmet } from 'react-helmet'
 import { BrowserRouter, Route, Switch } from 'react-router-dom'
 import 'regenerator-runtime/runtime'
-import styled, { ThemeProvider } from 'styled-components'
-import { Container as C } from '@material-ui/core'
+import { ThemeProvider } from 'styled-components'
 
 import { dimensions } from '@app/constants'
 import { AppBar, Drawer } from '@app/containers'
 import favicon from '@app/images/favicon.png'
 import { Home, Sura } from '@app/pages'
 import { base as baseTheme } from '@app/themes'
-
-const Container = styled(C)`
-  height: 100%;
-`
-
-const Page = styled.div`
-  display: flex;
-  flex: 1;
-  padding-left: ${(props) => props.paddingLeft}px;
-  padding-top: ${dimensions.APP_BAR_HEIGHT}px;
-`
-
-const Scrollbars = styled(S)`
-  flex: 1;
-`
+import { Container, Page, Scrollbars } from './styles'
 
 export default class Component extends React.Component {
   static propTypes = {
